make executor2 as the default executor
