+++ executor.py
@@ -1,9 +1,12 @@
 """
-Executor executes a script, parses consoles and posts console logs
+Executor executes a script, parses consoles and posts logs
 """
 
+from datetime import datetime
 import json
+import shutil
 import subprocess
+import tempfile
 import threading
 import time
 import traceback
@@ -13,20 +16,31 @@ from shippable_adapter import ShippableAdapter
 
 class Executor(object):
     """
-    Sets up config for the job, defaults for exit code and consoles
+    Sets up attributes that will be used to execute the steplet
     """
     def __init__(self, config):
         # -------
         # Private
         # -------
+
+        # Configs obtained from the job.env file
         self._config = config
         self._shippable_adapter = ShippableAdapter(config)
-        self._is_executing = False
 
-        # Consoles
-        # --------
-        self._console_buffer = []
-        self._console_buffer_lock = threading.Lock()
+        # Threads
+        self._logger_thread = None
+        self._script_runner_thread = None
+
+        # Error buffer state
+        self._has_errors = False
+
+        # Log directory and file
+        self._temporary_log_directory = tempfile.mkdtemp()
+        self._log_file_path = \
+            os.path.join(self._temporary_log_directory, 'logs')
+        buffer_size = 0
+        self._write_log_file = open(self._log_file_path, 'w', buffer_size)
+        self._read_log_file = open(self._log_file_path, 'r')
 
         # Console state
         self._current_group_info = None
@@ -34,7 +48,7 @@ class Executor(object):
         self._current_cmd_info = None
         self._show_group = None
 
-        # Errors
+        # Execution error consoles
         self._error_grp = {
             'consoleId': str(uuid.uuid4()),
             'parentConsoleId': 'root',
@@ -44,32 +58,38 @@ class Executor(object):
             'isSuccess': False
         }
         self._error_buffer = [self._error_grp]
-        self._has_errors = False
 
         # ------
         # Public
         # ------
+
+        # Assume failure by default
         self.exit_code = 1
+
+    def __del__(self):
+        shutil.rmtree(self._temporary_log_directory, ignore_errors=True)
 
     def execute(self):
         """
-        Starts threads to execute the script and flush consoles
+        Starts the script runner and logger threads and waits for
+        them to finish.
         """
-        script_runner_thread = threading.Thread(target=self._script_runner)
-        script_runner_thread.start()
 
-        # Wait for the execution to complete.
-        self._is_executing = True
-        console_flush_timer = threading.Timer(
-            self._config['CONSOLE_FLUSH_INTERVAL_SECONDS'],
-            self._set_console_flush_timer
-        )
-        console_flush_timer.start()
-        script_runner_thread.join()
-        self._is_executing = False
+        # Instantiate script runner and logger threads
+        self._script_runner_thread = \
+            threading.Thread(target=self._script_runner)
+        self._logger_thread = threading.Thread(target=self.logger)
+
+        # Start both the threads.
+        self._script_runner_thread.start()
+        self._logger_thread.start()
+
+        # Wait until the threads are completed
+        self._script_runner_thread.join()
+        self._logger_thread.join()
+
         if self._has_errors:
             self._flush_error_buffer()
-        self._flush_console_buffer()
 
     def _script_runner(self):
         """
@@ -77,12 +97,10 @@ class Executor(object):
         """
         # We need to unset the LD_LIBRARY_PATH set by pyinstaller. This
         # will ensure the script prefers libraries on system rather
-        # than the ones bundled during build time.
+        # than the ones bundled during steplet time.
         env = dict(os.environ)
         env.pop('LD_LIBRARY_PATH', None)
-
         cmd = self._config['SCRIPT_PATH']
-
         if self._config.get('REQEXEC_SHELL'):
             cmd = [self._config['REQEXEC_SHELL'], self._config['SCRIPT_PATH']]
 
@@ -91,7 +109,7 @@ class Executor(object):
                 cmd,
                 stdout=subprocess.PIPE,
                 stderr=subprocess.STDOUT,
-                cwd=self._config['BUILD_DIR'],
+                cwd=self._config['STEPLET_DIR'],
                 env=env
             )
         except Exception as ex:
@@ -116,9 +134,90 @@ class Executor(object):
 
         proc.kill()
 
+    def logger(self):
+        """
+        Reads from the log file and flushes consoles periodically or
+        if a limit is hit
+        """
+        logs_to_post = {
+            'stepletId': self._config['STEPLET_ID'],
+            'stepletConsoles': []
+        }
+
+        # Add a hidden version notice.
+        # NOTE: Remove this once we switch to this executor completely.
+        notice_console_id = str(uuid.uuid4())
+        notice_message = 'Notice: Executor v2'
+        logs_to_post['stepletConsoles'].append({
+            'consoleId': notice_console_id,
+            'parentConsoleId': 'root',
+            'type': 'grp',
+            'message': notice_message,
+            'timestamp': Executor._get_timestamp(),
+            'isShown': False
+        })
+
+        logs_to_post['stepletConsoles'].append({
+            'consoleId': notice_console_id,
+            'parentConsoleId': 'root',
+            'type': 'grp',
+            'message': notice_message,
+            'timestamp': Executor._get_timestamp(),
+            'timestampEndedAt': Executor._get_timestamp(),
+            'isSuccess': True,
+            'isShown': False
+        })
+
+        logs_last_posted_at = datetime.now()
+
+        while True:
+            post_logs = False
+            and_break = False
+
+            log_line = self._read_log_file.readline()
+
+            if log_line:
+                try:
+                    parsed_log_line = json.loads(log_line)
+                    logs_to_post['stepletConsoles'].append(parsed_log_line)
+                except Exception as ex:
+                    trace = traceback.format_exc()
+                    error = '{0}: {1}'.format(str(ex), trace)
+                    self._append_to_error_buffer(error)
+
+                # We added a new line, if the new array length exceeds max
+                # log lines, post logs.
+                if len(logs_to_post['stepletConsoles']) \
+                    >= self._config['MAX_LOG_LINES_TO_FLUSH']:
+                    post_logs = True
+            else:
+                # If the script runner is dead and there are no more logs to
+                # read, attempt to post any remaining logs and break.
+                if not self._script_runner_thread.isAlive():
+                    post_logs = True
+                    and_break = True
+                # If its been a while since we posted logs, post.
+                elif (datetime.now() - logs_last_posted_at).total_seconds() \
+                    > self._config['MAX_LOGS_FLUSH_WAIT_TIME_IN_S'] \
+                    and logs_to_post['stepletConsoles']:
+                    post_logs = True
+                # Sleep a bit if there hasn't been any activity.
+                else:
+                    time.sleep(self._config['LOGS_FILE_READ_WAIT_TIME_IN_S'])
+
+            # Post logs if asked and there is something to post.
+            if post_logs and logs_to_post['stepletConsoles']:
+                logs_last_posted_at = datetime.now()
+                data = json.dumps(logs_to_post)
+                self._shippable_adapter.post_step_let_consoles(data)
+                logs_to_post['stepletConsoles'] = []
+
+            if and_break:
+                break
+
     def _handle_console_line(self, line):
         """
-        Parses a single line of console output and pushes it to buffer
+        Parses a single line of console output and pushes it to a file
         This also returns whether the console line is successful and the
         script is complete
         """
@@ -141,7 +240,7 @@ class Executor(object):
                 'timestamp': timestamp,
                 'isShown': self._show_group
             }
-            self._append_to_console_buffer(console_out)
+            self._append_to_log_file(console_out)
         elif line.startswith('__SH__CMD__START__'):
             self._current_cmd_info = line_split[1]
             current_cmd_name = '|'.join(line_split[2:])
@@ -156,7 +255,7 @@ class Executor(object):
                 'timestamp': timestamp,
             }
             if parent_id:
-                self._append_to_console_buffer(console_out)
+                self._append_to_log_file(console_out)
         elif line.startswith('__SH__CMD__END__'):
             current_cmd_end_info = line_split[1]
             current_cmd_end_name = '|'.join(line_split[2:])
@@ -177,7 +276,7 @@ class Executor(object):
                 'isShown': self._show_group
             }
             if parent_id:
-                self._append_to_console_buffer(console_out)
+                self._append_to_log_file(console_out)
         elif line.startswith('__SH__GROUP__END__'):
             current_grp_end_info = line_split[1]
             current_grp_end_name = '|'.join(line_split[2:])
@@ -195,7 +294,7 @@ class Executor(object):
                 'isSuccess': is_cmd_success,
                 'isShown': self._show_group
             }
-            self._append_to_console_buffer(console_out)
+            self._append_to_log_file(console_out)
         elif line.startswith('__SH__SCRIPT_END_SUCCESS__'):
             is_script_success = True
             is_complete = True
@@ -213,73 +312,22 @@ class Executor(object):
                 'timestamp': timestamp,
             }
             if parent_id:
-                self._append_to_console_buffer(console_out)
+                self._append_to_log_file(console_out)
             else:
                 self._append_to_error_buffer(line)
 
         return is_script_success, is_complete
 
-    def _append_to_console_buffer(self, console_out):
+    def _append_to_log_file(self, console_out):
         """
-        Pushes a console line to buffer after taking over lock
+        Pushes a console line to the log file
         """
-        with self._console_buffer_lock:
-            self._console_buffer.append(console_out)
-
-        if len(self._console_buffer) > self._config['CONSOLE_BUFFER_LENGTH']:
-            self._flush_console_buffer()
-
-    def _set_console_flush_timer(self):
-        """
-        Calls _flush_console_buffer to flush console buffers in constant
-        intervals and stops when the script has finished execution
-        """
-        if not self._is_executing:
-            return
-
-        self._flush_console_buffer()
-        console_flush_timer = threading.Timer(
-            self._config['CONSOLE_FLUSH_INTERVAL_SECONDS'],
-            self._set_console_flush_timer
-        )
-        console_flush_timer.start()
-
-    def _flush_console_buffer(self):
-        """
-        Flushes console buffer after taking over lock
-        """
-        if self._console_buffer:
-            with self._console_buffer_lock:
-                # If there is an exception in stringifying the data, test
-                # each line to ensure only the sanitized ones are sent.
-                # Errors are pushed to the error buffer. Testing on failure
-                # will ensure that we don't test unnecessarily.
-                try:
-                    req_body = {
-                        'buildJobId': self._config['BUILD_JOB_ID'],
-                        'buildJobConsoles': self._console_buffer
-                    }
-                    json.dumps(req_body)
-                    data = json.dumps(req_body)
-                except Exception as ex:
-                    req_body = {
-                        'buildJobId': self._config['BUILD_JOB_ID'],
-                        'buildJobConsoles': []
-                    }
-
-                    for console in self._console_buffer:
-                        try:
-                            json.dumps(console)
-                            req_body['buildJobConsoles'].append(console)
-                        except Exception as ex:
-                            trace = traceback.format_exc()
-                            error = '{0}: {1}'.format(str(ex), trace)
-                            self._append_to_error_buffer(error)
-                    data = json.dumps(req_body)
-
-                self._shippable_adapter.post_build_job_consoles(data)
-                del self._console_buffer
-                self._console_buffer = []
+        try:
+            self._write_log_file.write(json.dumps(console_out) + '\n')
+        except Exception as ex:
+            trace = traceback.format_exc()
+            error = '{0}: {1}'.format(str(ex), trace)
+            self._append_to_error_buffer(error)
 
     def _append_to_error_buffer(self, error):
         """
@@ -304,11 +352,11 @@ class Executor(object):
         Flushes error buffer
         """
         req_body = {
-            'buildJobId': self._config['BUILD_JOB_ID'],
-            'buildJobConsoles': self._error_buffer
+            'stepletId': self._config['STEPLET_ID'],
+            'stepletConsoles': self._error_buffer
         }
         data = json.dumps(req_body)
-        self._shippable_adapter.post_build_job_consoles(data)
+        self._shippable_adapter.post_step_let_consoles(data)
         del self._error_buffer
         self._error_buffer = []
